Styled Portfolio Table / Looking to Add More Styling

+++ python-project-starter/react-app/src/components/Portfolio/PortfolioTableRow.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-
+import './PortfolioMain.css'
 
 const PortfolioTableRow = ({ s, setUpdate, user }) => {
     const [newAmount, setNewAmount] = useState(s.count)
@@ -21,6 +21,12 @@ const PortfolioTableRow = ({ s, setUpdate, user }) => {
     }
 
     const handleBuy = async () => {
+        if (newAmount === s.count) {
+            setEditBuy(false)
+            setEditContentBuy(false)
+            return
+        }
+
         const res = await fetch(`/api/portfolios/add/${s.stockId}/${user.id}`, {
             method: "PUT",
             headers: {
@@ -35,10 +41,17 @@ const PortfolioTableRow = ({ s, setUpdate, user }) => {
         setEditContentBuy(false)
     }
     const handleSell = async () => {
+        if (newAmount === s.count) {
+            setEditSell(false)
+            setEditContentSell(false)
+            return
+        }
+
         if (newAmount == 0) {
             const sell_res = await fetch(`/api/portfolios/sell/${s.stockId}/${user.id}`, {
                 method: "DELETE"
             })
+            setUpdate(true)
             setEditSell(false)
             setEditContentSell(false)
             return
@@ -53,6 +66,7 @@ const PortfolioTableRow = ({ s, setUpdate, user }) => {
                 "amount": parseFloat(newAmount)
             })
         })
+        setUpdate(true)
         setEditSell(false)
         setEditContentSell(false)
         return
@@ -70,28 +84,36 @@ const PortfolioTableRow = ({ s, setUpdate, user }) => {
     let btn;
 
     if (editBuy) {
-        btn = <td>
-        <button onClick={handleBuy}>Buy</button>
-        <button onClick={handleCancel}>Cancel</button>
+        btn = <td className="portfolio-btn-container">
+        <button className="green-btn" onClick={handleBuy}>Buy</button>
+        <button className="red-btn" onClick={handleCancel}>Cancel</button>
     </td>
     } else if (editSell) {
-        btn = <td>
-        <button onClick={handleSell}>Sell</button>
-        <button onClick={handleCancel}>Cancel</button>
+        btn = <td className="portfolio-btn-container">
+        <button className="red-btn" onClick={handleSell}>Sell</button>
+        <button className="red-btn" onClick={handleCancel}>Cancel</button>
     </td>
     } else {
-        btn = <td>
-                <button onClick={clickBuy}>Buy</button>
-                <button onClick={clickSell}>Sell</button>
+        btn = <td className="portfolio-btn-container">
+                <button className="green-btn" onClick={clickBuy}>Buy</button>
+                <button className="red-btn" onClick={clickSell}>Sell</button>
             </td>
     }
 
     let amountField
 
+    let profit
+
+    if ((parseFloat(s.currentPrice - s.purchasePrice) / s.purchasePrice * 100).toFixed(2) > 0) {
+        profit = <td id='portfolio-stock-profit'>{parseFloat((s.currentPrice - s.purchasePrice) / s.purchasePrice * 100).toFixed(2)}%</td>
+    } else {
+        profit = <td id='portfolio-stock-loss'>{parseFloat((s.currentPrice - s.purchasePrice) / s.purchasePrice * 100).toFixed(2)}%</td>
+    }
+
     if (editContentBuy) {
-        amountField = <td><input type="number" min={s.count} name="amount" value={newAmount} onChange={e => setNewAmount(e.target.value)}/></td>
+        amountField = <td><input id="edit-owned" type="number" min={s.count} name="amount" value={newAmount} onChange={e => setNewAmount(e.target.value)}/></td>
     } else if (editContentSell) {
-        amountField = <td><input type="number" min={0} max={s.count} name="amount" value={newAmount} onChange={e => setNewAmount(e.target.value)}/></td>
+        amountField = <td className="portfolio-td"><input type="number" id="edit-owned" min={0} max={s.count} name="amount" value={newAmount} onChange={e => setNewAmount(e.target.value)}/></td>
     } else {
         amountField = <td>{parseFloat(s.count).toFixed(3)}</td>
     }
@@ -103,7 +125,7 @@ const PortfolioTableRow = ({ s, setUpdate, user }) => {
             <td>${parseFloat(s.currentPrice).toFixed(2)}</td>
             {amountField}
             <td>${parseFloat(s.purchasePrice).toFixed(2)}</td>
-            <td>{parseFloat((s.currentPrice - s.purchasePrice) / s.purchasePrice * 100).toFixed(2)}%</td>
+            {profit}
             {btn}
         </tr>
     )
